Rename _index.py to _searchindexer.py.
