+++ signac/contrib/_searchindexer.py
